% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{leafletMapbox}
\alias{leafletMapbox}
\title{adds a Toolbar to draw shapes/points on the map}
\usage{
leafletMapbox(access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"), init = NULL,
  data = NULL, width = NULL, height = NULL, padding = 0,
  options = leafletOptions())
}
\arguments{
\item{access_token}{By default will look up for a Environment variable MAPBOX_ACCESS_TOKEN}

\item{init}{An optional id/url/tilejson for the basemap.}

\item{data}{the data for the map}

\item{width}{the width of the map}

\item{height}{the height of the map}

\item{padding}{the padding of the map}

\item{options}{the map options}

\item{map}{the map}
}
\value{
A HTML widget object, on which we can add graphics layers using
}
\description{
adds a Toolbar to draw shapes/points on the map
}

