% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\docType{data}
\name{mapbox.classicStyleIds}
\alias{mapbox.classicStyleIds}
\title{Map IDs for classic Mapbox styles}
\format{An object of class \code{list} of length 14.}
\usage{
mapbox.classicStyleIds
}
\description{
Map IDs for classic Mapbox styles
}
\keyword{datasets}

